+++ contrib/occtax/data/migration_2.0.0beta5toV2.sql
@@ -197,8 +197,8 @@ VALUES(
   releve.geom_4326,
   ST_CENTROID(releve.geom_4326),
   releve.geom_local,
-  (to_char(releve.date_min, 'DD/MM/YYYY') || ' ' || to_char(releve.date_min, 'hh:mm:ss'))::timestamp,
-  (to_char(releve.date_max, 'DD/MM/YYYY') || ' ' || to_char(releve.date_max, 'hh:mm:ss'))::timestamp,
+  (to_char(releve.date_min, 'DD/MM/YYYY') || ' ' || '00:00:00')::timestamp,
+  (to_char(releve.date_max, 'DD/MM/YYYY') || ' ' || '00:00:00')::timestamp,
   validation.validator_full_name,
   validation.validation_comment,
   COALESCE (myobservers.observers_name, releve.observers_txt),
@@ -214,10 +214,6 @@ END;
 $BODY$
   LANGUAGE plpgsql VOLATILE
   COST 100;
-
-
-
-
 
 
 ----------------------
@@ -350,7 +346,7 @@ BEGIN
     sample_number_proof = NEW.sample_number_proof,
     digital_proof = NEW.digital_proof,
     non_digital_proof = NEW.non_digital_proof,
-    comments  = CONCAT('Relevé : ',COALESCE(releve.comment, '-' ), ' Occurrence: ', COALESCE(NEW.comment, '-' )),
+    comments  = CONCAT('Relevé : ',COALESCE(releve.comment, '- ' ), ' Occurrence: ', COALESCE(NEW.comment, '-' )),
     last_action = 'U'
     WHERE unique_id_sinp IN (SELECT unique_id_sinp_occtax FROM pr_occtax.cor_counting_occtax WHERE id_occurrence_occtax = NEW.id_occurrence_occtax);
 
@@ -396,6 +392,7 @@ COST 100;
 
 
 -- UPDATE Releve
+-- UPDATE Releve
 CREATE OR REPLACE FUNCTION pr_occtax.fct_tri_synthese_update_releve()
   RETURNS trigger AS
 $BODY$
@@ -413,6 +410,8 @@ BEGIN
   ELSE 
     theobservers:= NEW.observers_txt;
   END IF;
+
+
   --mise à jour en synthese des informations correspondant au relevé uniquement
   UPDATE gn_synthese.synthese SET
       id_dataset = NEW.id_dataset,
@@ -420,8 +419,8 @@ BEGIN
       id_digitiser = NEW.id_digitiser,
       id_nomenclature_obs_technique = NEW.id_nomenclature_obs_technique,
       id_nomenclature_grp_typ = NEW.id_nomenclature_grp_typ,
-      date_min = (to_char(NEW.date_min, 'DD/MM/YYYY') || ' ' || COALESCE(to_char(NEW.hour_min, 'hh:mm:ss'), '00:00:00'))::timestamp,
-      date_max = (to_char(NEW.date_max, 'DD/MM/YYYY') || ' ' || COALESCE(to_char(NEW.hour_max, 'hh:mm:ss'), '00:00:00'))::timestamp,
+      date_min = (to_char(NEW.date_min, 'DD/MM/YYYY') || ' ' || '00:00:00')::timestamp,
+      date_max = (to_char(NEW.date_max, 'DD/MM/YYYY') || ' ' || '00:00:00')::timestamp,
       altitude_min = NEW.altitude_min,
       altitude_max = NEW.altitude_max,
       the_geom_4326 = NEW.geom_4326,
@@ -433,7 +432,7 @@ BEGIN
       FOR theoccurrence IN SELECT * FROM pr_occtax.t_occurrences_occtax WHERE id_releve_occtax = NEW.id_releve_occtax
       LOOP
           UPDATE gn_synthese.synthese SET
-                comments = CONCAT('Relevé: ',COALESCE(NEW.comment, '-'), 'Occurrence: ', COALESCE(theoccurrence.comment, '-'))
+                comments = CONCAT('Relevé: ',COALESCE(NEW.comment, '- '), 'Occurrence: ', COALESCE(theoccurrence.comment, '-'))
           WHERE unique_id_sinp IN (SELECT unnest(pr_occtax.get_unique_id_sinp_from_id_releve(NEW.id_releve_occtax::integer)));
       END LOOP;
   END IF;
@@ -442,9 +441,6 @@ END;
 $BODY$
   LANGUAGE plpgsql VOLATILE
   COST 100;
-
-
-
 
 
 -- suppression d'un relevé
@@ -467,7 +463,7 @@ CREATE OR REPLACE FUNCTION pr_occtax.fct_tri_synthese_insert_cor_role_releve()
   RETURNS trigger AS
 $BODY$
 DECLARE
-  uuids_counting integer;
+  uuids_counting uuid[];
 BEGIN
   -- récupération des id_counting à partir de l'id_releve
   SELECT INTO uuids_counting pr_occtax.get_unique_id_sinp_from_id_releve(NEW.id_releve_occtax::integer);
@@ -664,6 +660,182 @@ CREATE TRIGGER tri_delete_synthese_cor_role_releves_occtax
   FOR EACH ROW
   EXECUTE PROCEDURE pr_occtax.fct_tri_synthese_delete_cor_role_releve();
 
+---------
+--VIEWS--
+---------
+DROP VIEW IF EXISTS v_releve_occtax;
+CREATE OR REPLACE VIEW pr_occtax.v_releve_occtax AS
+ SELECT rel.id_releve_occtax,
+    rel.id_dataset,
+    rel.id_digitiser,
+    rel.date_min,
+    rel.date_max,
+    rel.altitude_min,
+    rel.altitude_max,
+    rel.meta_device_entry,
+    rel.comment,
+    rel.geom_4326,
+    rel."precision",
+    occ.id_occurrence_occtax,
+    occ.cd_nom,
+    occ.nom_cite,
+    t.lb_nom,
+    t.nom_valide,
+    t.nom_vern,
+    (((t.nom_complet_html::text || ' '::text) || rel.date_min::date) || '<br/>'::text) || string_agg(DISTINCT(obs.nom_role::text || ' '::text) || obs.prenom_role::text, ', '::text) AS leaflet_popup,
+    COALESCE ( string_agg(DISTINCT(obs.nom_role::text || ' '::text) || obs.prenom_role::text, ', '::text),rel.observers_txt) AS observateurs
+   FROM pr_occtax.t_releves_occtax rel
+     LEFT JOIN pr_occtax.t_occurrences_occtax occ ON rel.id_releve_occtax = occ.id_releve_occtax
+     LEFT JOIN taxonomie.taxref t ON occ.cd_nom = t.cd_nom
+     LEFT JOIN pr_occtax.cor_role_releves_occtax cor_role ON cor_role.id_releve_occtax = rel.id_releve_occtax
+     LEFT JOIN utilisateurs.t_roles obs ON cor_role.id_role = obs.id_role
+  GROUP BY rel.id_releve_occtax, rel.id_dataset, rel.id_digitiser, rel.date_min, rel.date_max, rel.altitude_min, rel.altitude_max, rel.meta_device_entry, rel.comment, rel.geom_4326, rel."precision", t.cd_nom, occ.nom_cite, occ.id_occurrence_occtax, t.lb_nom, t.nom_valide, t.nom_complet_html, t.nom_vern;
+
+--Vue représentant l'ensemble des relevés du protocole occtax pour la représentation du module carte liste
+CREATE OR REPLACE VIEW pr_occtax.v_releve_list AS
+ SELECT rel.id_releve_occtax,
+    rel.id_dataset,
+    rel.id_digitiser,
+    rel.date_min,
+    rel.date_max,
+    rel.altitude_min,
+    rel.altitude_max,
+    rel.meta_device_entry,
+    rel.comment,
+    rel.geom_4326,
+    rel."precision",
+   dataset.dataset_name,
+    string_agg(DISTINCT t.nom_valide::text, ','::text) AS taxons,
+    (((string_agg(DISTINCT t.nom_valide::text, ','::text) || '<br/>'::text) || rel.date_min::date) || '<br/>'::text) || COALESCE(string_agg(DISTINCT(obs.nom_role::text || ' '::text) || obs.prenom_role::text, ', '::text), rel.observers_txt::text) AS leaflet_popup,
+    COALESCE(string_agg(DISTINCT(obs.nom_role::text || ' '::text) || obs.prenom_role::text, ', '::text), rel.observers_txt::text) AS observateurs
+   FROM pr_occtax.t_releves_occtax rel
+     LEFT JOIN pr_occtax.t_occurrences_occtax occ ON rel.id_releve_occtax = occ.id_releve_occtax
+     LEFT JOIN taxonomie.taxref t ON occ.cd_nom = t.cd_nom
+     LEFT JOIN pr_occtax.cor_role_releves_occtax cor_role ON cor_role.id_releve_occtax = rel.id_releve_occtax
+     LEFT JOIN utilisateurs.t_roles obs ON cor_role.id_role = obs.id_role
+     LEFT JOIN gn_meta.t_datasets dataset ON dataset.id_dataset = rel.id_dataset
+  GROUP BY dataset.dataset_name, rel.id_releve_occtax, rel.id_dataset, rel.id_digitiser, rel.date_min, rel.date_max, rel.altitude_min, rel.altitude_max, rel.meta_device_entry;
+
+CREATE OR REPLACE VIEW pr_occtax.export_occtax_sinp AS 
+ SELECT ccc.unique_id_sinp_occtax AS "permId",
+    ref_nomenclatures.get_cd_nomenclature(occ.id_nomenclature_observation_status) AS "statObs",
+    occ.nom_cite AS "nomCite",
+    rel.date_min AS "dateDebut",
+    rel.date_max AS "dateFin",
+    rel.hour_min AS "heureDebut",
+    rel.hour_max AS "heureFin",
+    rel.altitude_max AS "altMax",
+    rel.altitude_min AS "altMin",
+    occ.cd_nom AS "cdNom",
+    taxonomie.find_cdref(occ.cd_nom) AS "cdRef",
+    gn_commons.get_default_parameter('taxref_version'::text, NULL::integer) AS "versionTAXREF",
+    rel.date_min AS datedet,
+    occ.comment,
+    'NSP'::text AS "dSPublique",
+    d.unique_dataset_id AS "jddMetadonneeDEEId",
+    ref_nomenclatures.get_cd_nomenclature(occ.id_nomenclature_source_status) AS "statSource",
+    '0'::text AS "diffusionNiveauPrecision",
+    ccc.unique_id_sinp_occtax AS "idOrigine",
+    d.dataset_name AS "jddCode",
+    d.unique_dataset_id AS "jddId",
+    NULL::text AS "refBiblio",
+    ref_nomenclatures.get_cd_nomenclature(occ.id_nomenclature_obs_meth) AS "obsMeth",
+    ref_nomenclatures.get_cd_nomenclature(occ.id_nomenclature_bio_condition) AS "ocEtatBio",
+    COALESCE(ref_nomenclatures.get_cd_nomenclature(occ.id_nomenclature_naturalness), '0'::text::character varying) AS "ocNat",
+    ref_nomenclatures.get_cd_nomenclature(ccc.id_nomenclature_sex) AS "ocSex",
+    ref_nomenclatures.get_cd_nomenclature(ccc.id_nomenclature_life_stage) AS "ocStade",
+    '0'::text AS "ocBiogeo",
+    COALESCE(ref_nomenclatures.get_cd_nomenclature(occ.id_nomenclature_bio_status), '0'::text::character varying) AS "ocStatBio",
+    COALESCE(ref_nomenclatures.get_cd_nomenclature(occ.id_nomenclature_exist_proof), '0'::text::character varying) AS "preuveOui",
+    ref_nomenclatures.get_nomenclature_label(occ.id_nomenclature_determination_method, 'fr'::character varying) AS "ocMethDet",
+    occ.digital_proof AS "preuvNum",
+    occ.non_digital_proof AS "preuvNoNum",
+    rel.comment AS "obsCtx",
+    rel.unique_id_sinp_grp AS "permIdGrp",
+    'Relevé'::text AS "methGrp",
+    'OBS'::text AS "typGrp",
+    ccc.count_max AS "denbrMax",
+    ccc.count_min AS "denbrMin",
+    ref_nomenclatures.get_cd_nomenclature(ccc.id_nomenclature_obj_count) AS "objDenbr",
+    ref_nomenclatures.get_cd_nomenclature(ccc.id_nomenclature_type_count) AS "typDenbr",
+    COALESCE(string_agg(DISTINCT (r.nom_role::text || ' '::text) || r.prenom_role::text, ','::text), rel.observers_txt::text) AS "obsId",
+    COALESCE(string_agg(DISTINCT r.organisme::text, ','::text), o.nom_organisme::text, 'NSP'::text) AS "obsNomOrg",
+    COALESCE(occ.determiner, 'Inconnu'::character varying) AS "detId",
+    'NSP'::text AS "detNomOrg",
+    'NSP'::text AS "orgGestDat",
+    st_astext(rel.geom_4326) AS "WKT",
+    'In'::text AS "natObjGeo"
+   FROM pr_occtax.t_releves_occtax rel
+     LEFT JOIN pr_occtax.t_occurrences_occtax occ ON rel.id_releve_occtax = occ.id_releve_occtax
+     LEFT JOIN pr_occtax.cor_counting_occtax ccc ON ccc.id_occurrence_occtax = occ.id_occurrence_occtax
+     LEFT JOIN taxonomie.taxref tax ON tax.cd_nom = occ.cd_nom
+     LEFT JOIN gn_meta.t_datasets d ON d.id_dataset = rel.id_dataset
+     LEFT JOIN pr_occtax.cor_role_releves_occtax cr ON cr.id_releve_occtax = rel.id_releve_occtax
+     LEFT JOIN utilisateurs.t_roles r ON r.id_role = cr.id_role
+     LEFT JOIN utilisateurs.bib_organismes o ON o.id_organisme = r.id_organisme
+  GROUP BY ccc.unique_id_sinp_occtax, d.unique_dataset_id, occ.id_nomenclature_bio_condition, occ.id_nomenclature_naturalness, ccc.id_nomenclature_sex, ccc.id_nomenclature_life_stage, occ.id_nomenclature_bio_status, occ.id_nomenclature_exist_proof, occ.id_nomenclature_determination_method, rel.unique_id_sinp_grp, d.id_nomenclature_source_status, occ.id_nomenclature_blurring, occ.id_nomenclature_diffusion_level, 'Pr'::text, occ.nom_cite, rel.date_min, rel.date_max, rel.hour_min, rel.hour_max, rel.altitude_max, rel.altitude_min, occ.cd_nom, occ.id_nomenclature_observation_status, (taxonomie.find_cdref(occ.cd_nom)), (gn_commons.get_default_parameter('taxref_version'::text, NULL::integer)), rel.comment, 'Ac'::text, rel.id_dataset, NULL::text, (ref_nomenclatures.get_cd_nomenclature(occ.id_nomenclature_source_status)), ccc.id_counting_occtax, d.dataset_name, occ.determiner, occ.comment, (ref_nomenclatures.get_cd_nomenclature(occ.id_nomenclature_obs_meth)), (ref_nomenclatures.get_cd_nomenclature(occ.id_nomenclature_bio_condition)), (COALESCE(ref_nomenclatures.get_cd_nomenclature(occ.id_nomenclature_naturalness), '0'::text::character varying)), (ref_nomenclatures.get_cd_nomenclature(ccc.id_nomenclature_sex)), (ref_nomenclatures.get_cd_nomenclature(ccc.id_nomenclature_life_stage)), '0'::text, (COALESCE(ref_nomenclatures.get_cd_nomenclature(occ.id_nomenclature_bio_status), '0'::text::character varying)), (COALESCE(ref_nomenclatures.get_cd_nomenclature(occ.id_nomenclature_exist_proof), '0'::text::character varying)), (ref_nomenclatures.get_nomenclature_label(occ.id_nomenclature_determination_method, 'fr'::character varying)), occ.digital_proof, occ.non_digital_proof, 'Relevé'::text, 'OBS'::text, ccc.count_max, ccc.count_min, (ref_nomenclatures.get_cd_nomenclature(ccc.id_nomenclature_obj_count)), (ref_nomenclatures.get_cd_nomenclature(ccc.id_nomenclature_type_count)), rel.observers_txt, 'NSP'::text, o.nom_organisme, (st_astext(rel.geom_4326)), 'In'::text;
+
+CREATE OR REPLACE VIEW pr_occtax.export_occtax_dlb AS 
+ SELECT ccc.unique_id_sinp_occtax AS "permId",
+    ref_nomenclatures.get_cd_nomenclature(occ.id_nomenclature_observation_status) AS "statObs",
+    occ.nom_cite AS "nomCite",
+    to_char(rel.date_min, 'DD/MM/YYYY'::text) AS "dateDebut",
+    to_char(rel.date_max, 'DD/MM/YYYY'::text) AS "dateFin",
+    rel.hour_min AS "heureDebut",
+    rel.hour_max AS "heureFin",
+    rel.altitude_max AS "altMax",
+    rel.altitude_min AS "altMin",
+    occ.cd_nom AS "cdNom",
+    taxonomie.find_cdref(occ.cd_nom) AS "cdRef",
+    to_char(rel.date_min, 'DD/MM/YYYY'::text) AS "dateDet",
+    occ.comment,
+    'NSP'::text AS "dSPublique",
+    ref_nomenclatures.get_cd_nomenclature(occ.id_nomenclature_source_status) AS "statSource",
+    ccc.unique_id_sinp_occtax AS "idOrigine",
+    d.unique_dataset_id AS "jddId",
+    NULL::text AS "refBiblio",
+    ref_nomenclatures.get_cd_nomenclature(occ.id_nomenclature_obs_meth) AS "obsMeth",
+    ref_nomenclatures.get_cd_nomenclature(occ.id_nomenclature_bio_condition) AS "ocEtatBio",
+    COALESCE(ref_nomenclatures.get_cd_nomenclature(occ.id_nomenclature_naturalness), '0'::text::character varying) AS "ocNat",
+    ref_nomenclatures.get_cd_nomenclature(ccc.id_nomenclature_sex) AS "ocSex",
+    ref_nomenclatures.get_cd_nomenclature(ccc.id_nomenclature_life_stage) AS "ocStade",
+    '0'::text AS "ocBiogeo",
+    COALESCE(ref_nomenclatures.get_cd_nomenclature(occ.id_nomenclature_bio_status), '0'::text::character varying) AS "ocStatBio",
+    COALESCE(ref_nomenclatures.get_cd_nomenclature(occ.id_nomenclature_exist_proof), '0'::text::character varying) AS "preuveOui",
+    ref_nomenclatures.get_cd_nomenclature(occ.id_nomenclature_determination_method) AS "ocMethDet",
+    occ.digital_proof AS "preuvNum",
+    occ.non_digital_proof AS "preuvNoNum",
+    rel.comment AS "obsCtx",
+    rel.unique_id_sinp_grp AS "permIdGrp",
+    'Relevé'::text AS "methGrp",
+    'OBS'::text AS "typGrp",
+    ccc.count_max AS "denbrMax",
+    ccc.count_min AS "denbrMin",
+    ref_nomenclatures.get_cd_nomenclature(ccc.id_nomenclature_obj_count) AS "objDenbr",
+    ref_nomenclatures.get_cd_nomenclature(ccc.id_nomenclature_type_count) AS "typDenbr",
+    COALESCE(string_agg(DISTINCT (r.nom_role::text || ' '::text) || r.prenom_role::text, ','::text), rel.observers_txt::text) AS "obsId",
+    COALESCE(string_agg(DISTINCT r.organisme::text, ','::text), o.nom_organisme::text, 'NSP'::text) AS "obsNomOrg",
+    COALESCE(occ.determiner, 'Inconnu'::character varying) AS "detId",
+    'NSP'::text AS "detNomOrg",
+    'NSP'::text AS "orgGestDat",
+    st_astext(rel.geom_4326) AS "WKT",
+    'In'::text AS "natObjGeo",
+    rel.date_min,
+    rel.date_max,
+    rel.id_dataset,
+    rel.id_releve_occtax,
+    occ.id_occurrence_occtax,
+    rel.id_digitiser,
+    rel.geom_4326
+   FROM pr_occtax.t_releves_occtax rel
+     LEFT JOIN pr_occtax.t_occurrences_occtax occ ON rel.id_releve_occtax = occ.id_releve_occtax
+     LEFT JOIN pr_occtax.cor_counting_occtax ccc ON ccc.id_occurrence_occtax = occ.id_occurrence_occtax
+     LEFT JOIN taxonomie.taxref tax ON tax.cd_nom = occ.cd_nom
+     LEFT JOIN gn_meta.t_datasets d ON d.id_dataset = rel.id_dataset
+     LEFT JOIN pr_occtax.cor_role_releves_occtax cr ON cr.id_releve_occtax = rel.id_releve_occtax
+     LEFT JOIN utilisateurs.t_roles r ON r.id_role = cr.id_role
+     LEFT JOIN utilisateurs.bib_organismes o ON o.id_organisme = r.id_organisme
+  GROUP BY rel.date_min, rel.date_max, rel.id_dataset, rel.unique_id_sinp_grp, occ.id_occurrence_occtax, rel.id_digitiser, rel.geom_4326, ccc.unique_id_sinp_occtax, d.unique_dataset_id, occ.id_nomenclature_bio_condition, occ.id_nomenclature_naturalness, ccc.id_nomenclature_sex, ccc.id_nomenclature_life_stage, occ.id_nomenclature_bio_status, occ.id_nomenclature_exist_proof, occ.id_nomenclature_determination_method, rel.id_releve_occtax, d.id_nomenclature_source_status, occ.id_nomenclature_blurring, occ.id_nomenclature_diffusion_level, 'Pr'::text, occ.nom_cite, rel.hour_min, rel.hour_max, rel.altitude_max, rel.altitude_min, occ.cd_nom, occ.id_nomenclature_observation_status, (taxonomie.find_cdref(occ.cd_nom)), rel.comment, 'Ac'::text, NULL::text, (ref_nomenclatures.get_cd_nomenclature(occ.id_nomenclature_source_status)), ccc.id_counting_occtax, d.dataset_name, occ.determiner, occ.comment, (ref_nomenclatures.get_cd_nomenclature(occ.id_nomenclature_obs_meth)), (ref_nomenclatures.get_cd_nomenclature(occ.id_nomenclature_bio_condition)), (COALESCE(ref_nomenclatures.get_cd_nomenclature(occ.id_nomenclature_naturalness), '0'::text::character varying)), (ref_nomenclatures.get_cd_nomenclature(ccc.id_nomenclature_sex)), (ref_nomenclatures.get_cd_nomenclature(ccc.id_nomenclature_life_stage)), '0'::text, (COALESCE(ref_nomenclatures.get_cd_nomenclature(occ.id_nomenclature_bio_status), '0'::text::character varying)), (COALESCE(ref_nomenclatures.get_cd_nomenclature(occ.id_nomenclature_exist_proof), '0'::text::character varying)), (ref_nomenclatures.get_cd_nomenclature(occ.id_nomenclature_determination_method)), occ.digital_proof, occ.non_digital_proof, 'Relevé'::text, 'OBS'::text, ccc.count_max, ccc.count_min, (ref_nomenclatures.get_cd_nomenclature(ccc.id_nomenclature_obj_count)), (ref_nomenclatures.get_cd_nomenclature(ccc.id_nomenclature_type_count)), rel.observers_txt, 'NSP'::text, o.nom_organisme, (st_astext(rel.geom_4326)), 'In'::text;
 
 --------------------
 -- ASSOCIATED DATA--