+++ docs/CHANGELOG.rst
@@ -2,7 +2,7 @@
 CHANGELOG
 =========
 
-2.0.0-rc.1 (unreleased)
+2.0.0-rc.1 (2018-09-21)
 -----------------------
 
 La version 2 de GeoNature est une refonte complète de l'application.
@@ -12,7 +12,7 @@ La version 2 de GeoNature est une refonte complète de l'application.
 * Refonte de la base de données pour la rendre plus standarde, plus générique et modulaire
 * Refonte ergonomique pour moderniser l'application
 
-Présentation et suivi du projet : https://github.com/PnX-SI/GeoNature/issues/168
+Présentation et suivi des développements : https://github.com/PnX-SI/GeoNature/issues/168
 
 **Accueil**
 
@@ -156,7 +156,7 @@ Pour les utilisateurs utilisant la version 2.0.0.beta5 :
         geonature generate_frontend_modules_route
         geonature frontend_build
 
-* Exécuter l'update de la BDD GeoNature (``data/migrations/2.0.0beta5to2.0.0.sql``) ainsi que celui du sous-module Nomenclature (https://github.com/PnX-SI/Nomenclature-api-module/blob/1.2.1/data/update1.1.0to1.2.1.sql)
+* Exécuter l'update de la BDD GeoNature (``data/migrations/2.0.0beta5to2.0.0.sql``), celui de Occtax (``contrib/occtax/data/migration_2.0.0beta5toV2.sql``) ainsi que celui du sous-module Nomenclature (https://github.com/PnX-SI/Nomenclature-api-module/blob/1.2.1/data/update1.1.0to1.2.1.sql)
 
 
 2.0.0.beta5 (2018-07-16)